feat: Add OSGi configuration for user info endpoints (#105)
Allow the connectors endpoints to be configured via OSGi configuration, instead of being hard-coded.
Note that the France Connect connector does not benefit from this change as its endpoints values are defined via a map, that contains the URL per environment. That would require a different refactoring (with changes in the JavaScript code) to properly externalize it.

The main benefit of this change is be able to change the endpoint URLs during the Cypress tests, so the URLs (Google, Facebook, etc.) can be mocked.